Sign In

+++ app/src/main/java/ru/netology/nework/dao/PostRemoteKeyDao.kt
@@ -10,10 +10,10 @@ import ru.netology.nework.entity.PostRemoteKeyEntity
 interface PostRemoteKeyDao {
 
     @Query("SELECT MAX('key') FROM PostRemoteKeyEntity")
-    suspend fun max(): Long?
+    suspend fun max(): Int?
 
     @Query("SELECT MIN('key') FROM PostRemoteKeyEntity")
-    suspend fun min(): Long?
+    suspend fun min(): Int?
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insert(postRemoteKeyEntity: PostRemoteKeyEntity)
